+++ src/components/nCheckboxInput/nCheckboxGroup.vue
@@ -12,7 +12,7 @@
 import nValidationAlerts from '@/utils/components/nValidationAlerts.vue'
 import formField from '@/utils/formField/index.js'
 import { messages } from '@/utils/validation'
-import { color, spacing } from '@/utils/styleVariables/variables'
+import { color, spacing } from '@/utils/styleVariables/helpers/variables'
 
 const customStyleVariables = [color, spacing]
