+++ src/components/nOverflowCarousel/nOverflowCarouselItem.vue
@@ -11,7 +11,7 @@
 
 <script>
 import styleVariables from '../../utils/styleVariables'
-import { width, height } from '../../utils/styleVariables/variables'
+import { width, height } from '../../utils/styleVariables/helpers/variables'
 
 const defaultStyleVariables = [width, height]
