+++ src/components/nRadioInput/nRadioInput.vue
@@ -26,7 +26,7 @@
 <script>
 import formField from '../../utils/formField/index.js'
 import nValidationAlerts from '../../utils/components/nValidationAlerts.vue'
-import { color } from '../../utils/styleVariables/variables'
+import { color } from '../../utils/styleVariables/helpers/variables'
 
 const customStyleVariables = [color]
