+++ src/utils/styleVariables/helpers/parseCamelToKebab.js
@@ -1,0 +1,1 @@
+export const parseCamelToKebab = (string) => string.replace(/\B(?:([A-Z])(?=[a-z]))|(?:(?<=[a-z0-9])([A-Z]))/g, '-$1$2').toLowerCase()
