+++ src/utils/styleVariables/index.js
@@ -1,6 +1,6 @@
-import { calculateCssSize } from './calculateCssSize'
+import { calculateCssSize } from './helpers/calculateCssSize'
+import { parseCamelToKebab } from './helpers/parseCamelToKebab'
 
-const camelToKebab = (string) => string.replace(/\B(?:([A-Z])(?=[a-z]))|(?:(?<=[a-z0-9])([A-Z]))/g, '-$1$2').toLowerCase()
 
 const fillStylingObject = ($$props, styleVariables) => {
   if (!styleVariables) return
@@ -13,7 +13,7 @@ const fillStylingObject = ($$props, styleVariables) => {
 
   let style = new Object()
   for (let styleVariable of styleVariables) {
-    const variableName = `--${camelToKebab(styleVariable.name)}`
+    const variableName = `--${parseCamelToKebab(styleVariable.name)}`
 
     style[variableName] = types[styleVariable.type]($$props[styleVariable.name])
   }
@@ -21,7 +21,7 @@ const fillStylingObject = ($$props, styleVariables) => {
   return style
 }
 
-export const styleVariables = (defaultStyleVariables, customStyleVariables = []) => ({
+const styleVariables = (defaultStyleVariables, customStyleVariables = []) => ({
   computed: {
     style () {
       const defaultVariables = fillStylingObject(this.$props, defaultStyleVariables)
